tests: Follow ignore file data module
Ignore-file editing now lives in data.ignore_files, and block/unblock
commands import it from that module.

The tests still describe those helpers as Git utility behavior. That keeps
the validation attached to the old utility boundary and leaves command
assertions importing the old path.

This commit continues the ignore-file boundary move by relocating the
.gitignore helper tests under tests/data and updating block/unblock command
tests to import path helpers from data.ignore_files.

The final commit will add an architecture guard that prevents ignore-file
helpers from returning to utils.git.
