+++ tests/commands/test_unblock_file.py
@@ -6,9 +6,9 @@ import pytest
 
 from git_stage_batch.commands.block_file import command_block_file
 from git_stage_batch.commands.unblock_file import command_unblock_file
+from git_stage_batch.data.ignore_files import get_gitignore_path, get_local_exclude_path
 from git_stage_batch.exceptions import CommandError
 from git_stage_batch.utils.file_io import append_file_path_to_file, read_file_paths_file
-from git_stage_batch.utils.git import get_gitignore_path, get_local_exclude_path
 from git_stage_batch.utils.paths import get_blocked_files_file_path
